ml_service/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
python-dotenv==1.0.0
requests==2.31.0
exifread==3.0.0
